requirements.txt
mysql-connector-python
python-dotenv
pandas
paramiko